+++ src/input/aac_framing_packet_converter.h
@@ -24,7 +24,7 @@ protected:
   aac::parser_c m_parser;
 
 public:
-  aac_framing_packet_converter_c(generic_packetizer_c *ptzr);
+  aac_framing_packet_converter_c(generic_packetizer_c *ptzr, aac::parser_c::multiplex_type_e multiplex_type);
   virtual ~aac_framing_packet_converter_c() {};
 
   virtual bool convert(packet_cptr const &packet);